pyproject.toml
[tool.poetry]
name = "aocbot"
version = "0.1.0"
description = ""
authors = ["celeste <parafactual@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
openai = "^0.28.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
